% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{plot_scores}
\alias{plot_scores}
\title{Plot Multivariate Sparklines}
\usage{
plot_scores(
  scores,
  twogoalline = FALSE,
  nozerodots = FALSE,
  outlined = FALSE,
  show = TRUE,
  output_path = NULL
)
}
\arguments{
\item{scores}{A list, a list of lists, a data.frame or a list of data.frames
of match scores. Each match score has the form of a list (or a data.frame
row) of (home team score: int, away team score: int, is away game: logical)}

\item{twogoalline}{A logical, draws lines for two-goal levels,
by default FALSE.}

\item{nozerodots}{A logical, removes the dots placed for zero scores,
by default FALSE.}

\item{outlined}{A logical, only plots the outlines of the sparklines,
by default FALSE.}

\item{show}{A logical, prints the plot, by default TRUE.}

\item{output_path}{A string, Path to save the plot at (image type
is inferred from the path), by default NULL.}
}
\value{
ggplot objects of the plots created.
}
\description{
Plot the given list or lists of match results as multivariate sparklines.
This function can either show the plot, or save it to a given path.
}
\examples{
plot_scores(list(
     list(1,2,T), list(3,3,F), list(0,2,T), list(0,0,F), list(6,6,T)
))
plot_scores(list(
     list(1,2,T), list(3,3,F), list(0,2,T), list(0,0,F), list(6,6,T),
     list(0,2,F), list(0,0,F), list(6,6,T), list(6,3,T)
), outlined=T, output_path = "out.png")
}
