% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{plot_scores}
\alias{plot_scores}
\title{Plot Multivariate Sparklines}
\usage{
plot_scores(
  scores,
  twogoalline = FALSE,
  nozerodots = FALSE,
  outlined = FALSE,
  show = TRUE,
  output_path = NULL,
  ...
)
}
\arguments{
\item{scores}{A list, a list of lists, a data.frame or a list of data.frames
of match scores. Each match score has the form of a list (or a data.frame
row) of (home team score: int, away team score: int, is away game: logical)}

\item{twogoalline}{A logical, draws lines for two-goal levels,
by default FALSE.}

\item{nozerodots}{A logical, removes the dots placed for zero scores,
by default FALSE.}

\item{outlined}{A logical, only plots the outlines of the sparklines,
by default FALSE.}

\item{show}{A logical, prints the plot, by default TRUE.}

\item{output_path}{A string, Path to save the plot at (image type
is inferred from the path), by default NULL.}

\item{...}{A named list, additional configuration keywords for the
visualization. Not necessarily consistent with its latex counterpart,
but mostly a superset of it.
\itemize{
\item figure_height: Figure height in inches, by default 4
\item figure_width_per_match: Figure width per match in inches, by default 0.5
\item dpi: Dots per inch resolution, by default 300
\item thickness: Line thickness in cartesian coordinates, by default 0.18
\item edge_thickness: Edge thickness for outlined away games
(when outlined = True), by default 3
\item zerodot: Zero-dot radius ratio to thickness (when nozerodots = False),
by default 0.4
\item slant: Slope for unbalanced scores in degrees, by default 14
\item spacing: Spacing between matches in cartesian coordinates, by default 0.9
\item padding: Padding before and after the matches in cartesian coordinates,
by default 0.9
\item baseline_factor: Thickness of baseline with respect to line thickness,
by default 0.2
\item brighten: Brightness percentage of the two-goal lines (when
twogoalline = True) and away games (when outlined = False),
by default 33
\item transparent_background: Set the background transparent instead of white,
by default False
\item home_color: Color of home match lines in any of the three kinds of R
color specifications (i.e. either a color name as listed by colors(),
a hexadecimal string of the form "#rrggbb" or "#rrggbbaa",
or a positive integer i meaning palette()\link{i}), by default "black"
\item away_color: Color of away match lines in any of the three kinds of R
color specifications, by default "black"
\item baseline_color: Color of baselines in any of the three kinds of R
color specifications, by default "black"
\item fill_color: Fill color for the outlined sparklines in any of the three
kinds of R color specifications, by default NA.
\item clip_slanted_lines: Clip the ends of the slanted lines, by default True
}}
}
\value{
ggplot objects of the plots created.
}
\description{
Plot the given list or lists of match results as multivariate sparklines.
This function can either show the plot, or save it to a given path.
}
\examples{
plot_scores(list(
    list(1, 2, TRUE), list(3, 3, FALSE), list(0, 2, TRUE),
    list(0, 0, FALSE), list(6, 6, TRUE)
))
plot_scores(list(
    list(1, 2, TRUE), list(3, 3, FALSE), list(0, 2, TRUE), list(0, 0, FALSE),
    list(6, 6, TRUE), list(0, 2, FALSE), list(0, 0, FALSE), list(6, 6, TRUE),
    list(6, 3, TRUE)
), outlined = TRUE, output_path = "out.png")
plot_scores(list(
    list(8, 0, FALSE), list(4, 1, TRUE), list(4, 4, FALSE), list(1, 4, TRUE),
    list(5, 0, FALSE), list(0, 0, TRUE), list(1, 1, FALSE), list(2, 3, TRUE),
    list(NA, NA, FALSE), list(NA, NA, TRUE), list(NA, NA, FALSE)
), home_color = "red")
}
